Stop the search for wscript if the directory contains the waf script as well
If jack2 is a git submodule dir and the parent dir contains wscripts as well,
the waf exectued from withing the jack2 directory prefers the wscript
from the parent directory. When the waf in the parent directory is
newer one, the older waf tries to load tools that are not available.

In particular this situation happens in ladish. ladish waf-1.5.18 has
jack2 as git submodule. jack2 has waf-1.5.0. waf-1.5.0 in the jack2
subdirectory loads the wscript in the ladish directory and which in
turn tries to load the boost tool that is not available in waf-1.5.0

The replacement waf script is slighly modified waf-1.5.0