Split ADDON_DIR into JACK_DRIVER_DIR & JACK_INTERNAL_DIR and deprecate it.
It is better to have internal clients and drivers in different directories.

While prefixing the shared library names with "jack_" could be used for
distinguising between drivers and internal clients during load,
this was is not the case with current code.

Also, as neither drivers nor internal clients are supposed to be implicitly
loadable via ld.so as regular shared libraries, they now reside in

<PREFIX>/libexec/jack-driver/ and <PREFIX>/libexec/jack-internal/

The jack_ prefix for drivers is removed.

Windows specific code paths are not updated.