common/jack/session.h: remove newsm recommendation
Jack-Session was bad idea from the start,
so its use was never encouraged in LADI project.

NonSM & NewSM OSC-RPC, as used over UDP only, was and still is
unrealiable in multi computer in a network scenarios because of the
UDP packet loss.

Jack-Session used to be another jackaudio.org recommendation,
which ended in deprecations. However, as loading JS sessions via
LADISH is a feature to be maintained and sensible warnings during
build are to be easy catched, the half-deprecated jack-session api is
now undeprecated for the compiler. Jack-Session is left documented as
deprecated.